fix: line 38: middleware was using component-dependent stuff
