+++ pages/evaluation/initialisation/[assessmentId]/questions-objectives/index.vue
@@ -1,39 +1,41 @@
 <template>
-  <PageSection
-    :title="pageStore.evaluationInitiationPage.objectiveQuestionsTitle"
-    :intro="pageStore.evaluationInitiationPage.objectiveQuestionsDescription"
-    :is-first-element="true"
-    :intro-is-rich-text="true"
-    class="column is-8 questionnaire-container"
-  >
-    <div class="nav-questionnaire-container">
-      <button
-        class="button is-shade-600 is-rounded mt-4"
-        @click.prevent="goToFirstObjectiveQuestion"
-      >
-        <span>{{
-          pageStore.evaluationInitiationPage.objectiveQuestionsCallToAction
-        }}</span>
-        <span class="icon">
-          <icon
-            v-if="isLoading"
-            size="20"
-            name="loader-2-line"
-          />
-          <icon
-            v-else
-            size="20"
-            name="arrow-right-line"
-          />
-        </span>
-      </button>
-      <span
-        v-if="isLoading"
-        class="is-size-7 has-text-shade-600"
-      >{{ $t("en cours de chargement") }}</span>
-      <!-- <ButtonsArrowButton class="arrow-button-fixed is-left" @click.prevent="goBack" color="no-pillar" /> -->
-    </div>
-  </PageSection>
+  <NeedsParticipationOrAssessment>
+    <PageSection
+      :title="pageStore.evaluationInitiationPage.objectiveQuestionsTitle"
+      :intro="pageStore.evaluationInitiationPage.objectiveQuestionsDescription"
+      :is-first-element="true"
+      :intro-is-rich-text="true"
+      class="column is-8 questionnaire-container"
+    >
+      <div class="nav-questionnaire-container">
+        <button
+          class="button is-shade-600 is-rounded mt-4"
+          @click.prevent="goToFirstObjectiveQuestion"
+        >
+          <span>{{
+            pageStore.evaluationInitiationPage.objectiveQuestionsCallToAction
+          }}</span>
+          <span class="icon">
+            <icon
+              v-if="isLoading"
+              size="20"
+              name="loader-2-line"
+            />
+            <icon
+              v-else
+              size="20"
+              name="arrow-right-line"
+            />
+          </span>
+        </button>
+        <span
+          v-if="isLoading"
+          class="is-size-7 has-text-shade-600"
+        >{{ $t("en cours de chargement") }}</span>
+        <!-- <ButtonsArrowButton class="arrow-button-fixed is-left" @click.prevent="goBack" color="no-pillar" /> -->
+      </div>
+    </PageSection>
+  </NeedsParticipationOrAssessment>
 </template>
 
 <script setup lang="ts">
