+++ pages/evaluation/participation/[assessmentId]/role.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="container is-tight">
+  <NeedsParticipationOrAssessment class="container is-tight">
     <h1 class="title is-3 mt-2 has-text-black">
       {{ pageStore.evaluationQuestionnairePage.roleQuestionTitle }}
     </h1>
@@ -20,7 +20,7 @@
       :is-loading="isLoading"
       @submit="onSubmit"
     />
-  </div>
+  </NeedsParticipationOrAssessment>
 </template>
 
 <script setup lang="ts">
